% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipums_clean_dta.R
\name{ipums_clean_dta}
\alias{ipums_clean_dta}
\title{Clean IPUMS International Data Imported from Stata File}
\usage{
ipums_clean_dta(
  d = NULL,
  numeric_convert = NULL,
  string_as_factors = TRUE,
  clean_labels = FALSE,
  mis_lab = NULL
)
}
\arguments{
\item{d}{Data frame with labels, as typically obtained when imported using haven::read_dta()}

\item{numeric_convert}{Vector of character strings for columns to convert to numeric types}

\item{string_as_factors}{Covert any factors to strings.}

\item{clean_labels}{Clean labels. Extra step, often not required.}

\item{mis_lab}{Vector of character strings to convert labels for missing values to `NA`}
}
\value{
Cleaned data frame.
}
\description{
Clean IPUMS International Data Imported from Stata File
}
