% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_model_image.R
\name{tab_model_image}
\alias{tab_model_image}
\title{Convert a tab_model to an image}
\usage{
tab_model_image(..., file = "./image-table/tab", zoom = 1)
}
\arguments{
\item{...}{Passed to `sjPlot::tab_model`}

\item{file}{Character string. The file name of the output image. The file extension will be added automatically.}

\item{zoom}{Numeric. The zoom factor for the image. Default is 1.}
}
\value{
A table image saved as a PNG file.
}
\description{
Convert a tab_model to an image
}
